requirements.txt
asgiref==3.4.1
beautifulsoup4==4.11.1
certifi==2022.5.18
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
distlib==0.3.6
Django==4.0
django-cors-headers==3.12.0
djangorestframework==3.13.1
filelock==3.10.2
fiscalyear==0.4.0
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.24.1
pandas==1.5.3
platformdirs==3.1.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
virtualenv==20.21.0
Werkzeug==2.0.2
